Extract EngineDownloader to own file
Extract the EngineDownloader from the InstallEngineMojo as preparation
for multiple downloader services (URL based downloader, maven based
downloader...)
